+++ tests/testthat/test-summary_methods.R
@@ -303,8 +303,8 @@ test_that("getGrowthCurves works with MizerSim", {
 # summary ----
 test_that("summary works", {
     # Check that it works also with nonstandard kernel
-    params@species_params$ppmr_min = 100
-    params@species_params$ppmr_max = 10000
+    params@species_params$ppmr_min <- 100
+    params@species_params$ppmr_max <- 10000
     params@species_params$beta <- NULL
     params@species_params$sigma <- NULL
     species_params(params)$pred_kernel_type <- "box"